chore: Update SuperAdminDashboard component
