+++ src/pages/SuperAdmin/ViewAll.jsx
@@ -9,6 +9,7 @@ import { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 
 
+
 export default function SuperAdminDashboard() {
   const base_url = import.meta.env.VITE_APP_BACKEND_URL;
